Moved Colin's code changes onto head
I did a command : git checkout 'master@{2015-02-25 16:00:00}' to get
colin's code and then copied all the files to temp and then did a used
github desktop to get back to head then pulled files colin changed to
head.  This is a hack, we really need to be using branches.  Maybe a
branch for each developer for when developing at home.  Or maybe more
just always sync before and after do work to avoid overwriting a whole
set of changes like we had here.