Added some binary decode code to see status flag
Buried in the returned Diagnosis is "COF (Cordic Overflow), logic high
indicates an out of range error in the CORDIC part. When this bit is
set, the angle and
magnitude data is invalid. The absolute output maintains the last valid
angular value." so this is some code pulled from the internet to help
access that flag